Took Nate's concerns to heart and gave it a second run through
